Unbundle decode configuration (#275)
Rework the decode configuration so it is flags, not modes. The decode checks that make decode modes like CDE and Preferred and dCBOR are available individually.


* Unbundle decode configuration

* straggler

* fix test fan out

* Documentation updates

---------

+++ inc/qcbor/qcbor_private.h
@@ -256,7 +256,7 @@ struct _QCBOREncodeContext {
  * 32-bit machine size is 200 bytes
  */
 typedef struct __QCBORDecodeNesting  {
-  /* PRIVATE DATA STRUCTURE */
+   /* PRIVATE DATA STRUCTURE */
    struct nesting_decode_level {
       /*
        * This keeps tracking info for each nesting level. There are two
@@ -346,6 +346,7 @@ typedef struct  {
 typedef uint16_t QCBORMappedTagNumbers[QCBOR_MAX_TAGS_PER_ITEM1];
 
 
+
 /*
  * PRIVATE DATA STRUCTURE
  *
@@ -380,7 +381,7 @@ struct _QCBORDecodeContext {
 #define QCBOR_MAP_OFFSET_CACHE_INVALID UINT32_MAX
    uint32_t uMapEndOffsetCache;
 
-   uint8_t  uDecodeMode;
+   uint32_t uDecodeMode;
    uint8_t  bStringAllocateAll;
    uint8_t  uLastError;  /* QCBORError stuffed into a uint8_t */
    uint8_t  bAllowAllLabels; /* Used internally only, not an external feature yet */
